cd fix

+++ Program.cs
@@ -132,8 +132,11 @@ namespace realclicore
                         if (slnProj == 0 && slnData == 0 && slnDomain == 0 && slnBusiness == 0)
                         {
 
-                            await RunCli("cd..");
-                            await RunCli("rm -r realcorecli/*");
+                            var e = await RunCli("cd ..");
+                            if(e==0) {
+                                await RunCli("rm -r realcorecli/*");
+                            }
+                            
                         
                         }
                     }
@@ -144,8 +147,10 @@ namespace realclicore
                 var app = await RunCli(cli + " --output ..//" + projectname);
                 if (app == 0)
                 {
-                    await RunCli("cd..");
-                    await RunCli("rm -r realcorecli/*");
+                    var e = await RunCli("cd ..");
+                    if(e==0) {
+                        await RunCli("rm -r realcorecli/*");
+                    }
                     
                 }
             }
